[BuildScripts] Refactor the build scripts.
The old build scripts are un-human to read, and to maintain it. Therefore in this refactor work I moved all of the dependencies onto QiNiu CDN, which plays a mirror service role.
And the name of each dependency are all in the form of XXXX.Major.Minor.X, downloaded as  a .zip or tar.gz. Each of the zip file contains the directory of the name.
Through this approach we can easily manage the dependencies.