[Raft] Fix bug in pre-campaign process.
Problem: During re-election, the 2 active raft nodes (3-node-cluster) failed
to elect a leader. None of the nodes accepted the prevote requests from the other,
which makes the cluster became unavailable.

Reason: Each of the node will reset its election timer to 0 when they start
a pre-campaign. As the result, the 2 pre-candidates reject each other
since the election timer was not expired.

Solution: reset current leader to 0 when a node becomes pre-candidate.