[Raft] introduce & fix tests of raft snapshot and progress control
- We refactored the network simulator in this commit. It's much simpler
than the previous design. It was designed to be a low-level facility, which
allows the tester to ensure that messages are going as they want.
Now we wrap up the processes of log replication and election. The messages going
are opaque to the testers.
It's difficult to manage all of the messages sequence in a bunch of tests, especially
after we introduced progress control.

- We don't drop the msg when its term is not set. It's convenient that
local message like MsgBeat, MsgProp do not have to set term. It used to be
unsafe, because vote or app with unset term will be permitted. Now the code
is quite fixed, so there's no need to worry.

- Some tests failure were fixed.